requirements.txt

Django==4.2.9
djangorestframework==3.14.0
imgkit==1.2.3
xhtml2pdf==0.2.13
